Move pixelwise
